improve: extract files in cpp/hpp and precompiled header option
Change solution for auto include protobuf generated
