+++ cmake/modules/CanaryLib.cmake
@@ -23,7 +23,13 @@ add_subdirectory(utils)
 target_sources(${PROJECT_NAME}_lib PRIVATE canary_server.cpp)
 
 # Add public pre compiler header to lib, to pass down to related targets
-target_precompile_headers(${PROJECT_NAME}_lib PUBLIC pch.hpp)
+if (NOT SPEED_UP_BUILD_UNITY)
+    target_precompile_headers(${PROJECT_NAME}_lib PUBLIC pch.hpp)
+endif()
+
+if(NOT SPEED_UP_BUILD_UNITY AND USE_PRECOMPILED_HEADERS)
+    target_compile_definitions(${PROJECT_NAME}_lib PUBLIC -DUSE_PRECOMPILED_HEADER)
+endif()
 
 # *****************************************************************************
 # Build flags - need to be set before the links and sources
