+++ src/creatures/creatures_definitions.hpp
@@ -9,8 +9,18 @@
 
 #pragma once
 
-// Enum
+#ifndef USE_PRECOMPILED_HEADERS
+	#include <string>
+	#include <vector>
+	#include <map>
+	#include <list>
+	#include <utility>
+	#include <cstdint>
+	#include <memory>
+	#include <cmath>
+#endif
 
+// Enum
 enum SkillsId_t {
 	SKILLVALUE_LEVEL = 0,
 	SKILLVALUE_TRIES = 1,
@@ -316,72 +326,6 @@ enum MarketOfferState_t {
 
 	OFFERSTATE_ACCEPTEDEX = 255,
 };
-
-enum ObjectCategory_t {
-	OBJECTCATEGORY_NONE = 0,
-	OBJECTCATEGORY_ARMORS = 1,
-	OBJECTCATEGORY_NECKLACES = 2,
-	OBJECTCATEGORY_BOOTS = 3,
-	OBJECTCATEGORY_CONTAINERS = 4,
-	OBJECTCATEGORY_DECORATION = 5,
-	OBJECTCATEGORY_FOOD = 6,
-	OBJECTCATEGORY_HELMETS = 7,
-	OBJECTCATEGORY_LEGS = 8,
-	OBJECTCATEGORY_OTHERS = 9,
-	OBJECTCATEGORY_POTIONS = 10,
-	OBJECTCATEGORY_RINGS = 11,
-	OBJECTCATEGORY_RUNES = 12,
-	OBJECTCATEGORY_SHIELDS = 13,
-	OBJECTCATEGORY_TOOLS = 14,
-	OBJECTCATEGORY_VALUABLES = 15,
-	OBJECTCATEGORY_AMMO = 16,
-	OBJECTCATEGORY_AXES = 17,
-	OBJECTCATEGORY_CLUBS = 18,
-	OBJECTCATEGORY_DISTANCEWEAPONS = 19,
-	OBJECTCATEGORY_SWORDS = 20,
-	OBJECTCATEGORY_WANDS = 21,
-	OBJECTCATEGORY_PREMIUMSCROLLS = 22, // not used in quickloot
-	OBJECTCATEGORY_TIBIACOINS = 23, // not used in quickloot
-	OBJECTCATEGORY_CREATUREPRODUCTS = 24,
-	OBJECTCATEGORY_GOLD = 30,
-	OBJECTCATEGORY_DEFAULT = 31, // unassigned loot
-
-	OBJECTCATEGORY_FIRST = OBJECTCATEGORY_ARMORS,
-	OBJECTCATEGORY_LAST = OBJECTCATEGORY_DEFAULT,
-};
-
-static bool isValidObjectCategory(uint8_t category) {
-	static std::unordered_set<uint8_t> valid = {
-		OBJECTCATEGORY_NONE,
-		OBJECTCATEGORY_ARMORS,
-		OBJECTCATEGORY_NECKLACES,
-		OBJECTCATEGORY_BOOTS,
-		OBJECTCATEGORY_CONTAINERS,
-		OBJECTCATEGORY_DECORATION,
-		OBJECTCATEGORY_FOOD,
-		OBJECTCATEGORY_HELMETS,
-		OBJECTCATEGORY_LEGS,
-		OBJECTCATEGORY_OTHERS,
-		OBJECTCATEGORY_POTIONS,
-		OBJECTCATEGORY_RINGS,
-		OBJECTCATEGORY_RUNES,
-		OBJECTCATEGORY_SHIELDS,
-		OBJECTCATEGORY_TOOLS,
-		OBJECTCATEGORY_VALUABLES,
-		OBJECTCATEGORY_AMMO,
-		OBJECTCATEGORY_AXES,
-		OBJECTCATEGORY_CLUBS,
-		OBJECTCATEGORY_DISTANCEWEAPONS,
-		OBJECTCATEGORY_SWORDS,
-		OBJECTCATEGORY_WANDS,
-		OBJECTCATEGORY_PREMIUMSCROLLS,
-		OBJECTCATEGORY_TIBIACOINS,
-		OBJECTCATEGORY_CREATUREPRODUCTS,
-		OBJECTCATEGORY_GOLD,
-		OBJECTCATEGORY_DEFAULT,
-	};
-	return valid.contains(category);
-}
 
 enum RespawnPeriod_t {
 	RESPAWNPERIOD_ALL,
@@ -1488,7 +1432,7 @@ struct FamiliarEntry {
 struct Skill {
 	uint64_t tries = 0;
 	uint16_t level = 10;
-	double_t percent = 0;
+	double percent = 0;
 };
 
 struct Kill {
