+++ src/database/database.hpp
@@ -12,6 +12,11 @@
 #include "declarations.hpp"
 #include "lib/logging/log_with_spd_log.hpp"
 
+#ifndef USE_PRECOMPILED_HEADERS
+	#include <mysql/mysql.h>
+	#include <mutex>
+#endif
+
 class DBResult;
 using DBResult_ptr = std::shared_ptr<DBResult>;
 
@@ -58,9 +63,7 @@ private:
 	bool rollback();
 	bool commit();
 
-	bool isRecoverableError(unsigned int error) const {
-		return error == CR_SERVER_LOST || error == CR_SERVER_GONE_ERROR || error == CR_CONN_HOST_ERROR || error == 1053 /*ER_SERVER_SHUTDOWN*/ || error == CR_CONNECTION_ERROR;
-	}
+	bool isRecoverableError(unsigned int error) const;
 
 	MYSQL* handle = nullptr;
 	std::recursive_mutex databaseLock;
