+++ src/enums/lua_variant_type.hpp
@@ -1,0 +1,19 @@
+/**
+ * Canary - A free and open-source MMORPG server emulator
+ * Copyright (©) 2019-2022 OpenTibiaBR <opentibiabr@outlook.com>
+ * Repository: https://github.com/opentibiabr/canary
+ * License: https://github.com/opentibiabr/canary/blob/main/LICENSE
+ * Contributors: https://github.com/opentibiabr/canary/graphs/contributors
+ * Website: https://docs.opentibiabr.com/
+ */
+
+#pragma once
+
+enum LuaVariantType_t {
+	VARIANT_NONE,
+
+	VARIANT_NUMBER,
+	VARIANT_POSITION,
+	VARIANT_TARGETPOSITION,
+	VARIANT_STRING,
+};
