+++ src/enums/object_category.hpp
@@ -1,0 +1,43 @@
+/**
+ * Canary - A free and open-source MMORPG server emulator
+ * Copyright (©) 2019-2022 OpenTibiaBR <opentibiabr@outlook.com>
+ * Repository: https://github.com/opentibiabr/canary
+ * License: https://github.com/opentibiabr/canary/blob/main/LICENSE
+ * Contributors: https://github.com/opentibiabr/canary/graphs/contributors
+ * Website: https://docs.opentibiabr.com/
+ */
+
+#pragma once
+
+enum ObjectCategory_t {
+	OBJECTCATEGORY_NONE = 0,
+	OBJECTCATEGORY_ARMORS = 1,
+	OBJECTCATEGORY_NECKLACES = 2,
+	OBJECTCATEGORY_BOOTS = 3,
+	OBJECTCATEGORY_CONTAINERS = 4,
+	OBJECTCATEGORY_DECORATION = 5,
+	OBJECTCATEGORY_FOOD = 6,
+	OBJECTCATEGORY_HELMETS = 7,
+	OBJECTCATEGORY_LEGS = 8,
+	OBJECTCATEGORY_OTHERS = 9,
+	OBJECTCATEGORY_POTIONS = 10,
+	OBJECTCATEGORY_RINGS = 11,
+	OBJECTCATEGORY_RUNES = 12,
+	OBJECTCATEGORY_SHIELDS = 13,
+	OBJECTCATEGORY_TOOLS = 14,
+	OBJECTCATEGORY_VALUABLES = 15,
+	OBJECTCATEGORY_AMMO = 16,
+	OBJECTCATEGORY_AXES = 17,
+	OBJECTCATEGORY_CLUBS = 18,
+	OBJECTCATEGORY_DISTANCEWEAPONS = 19,
+	OBJECTCATEGORY_SWORDS = 20,
+	OBJECTCATEGORY_WANDS = 21,
+	OBJECTCATEGORY_PREMIUMSCROLLS = 22, // not used in quickloot
+	OBJECTCATEGORY_TIBIACOINS = 23, // not used in quickloot
+	OBJECTCATEGORY_CREATUREPRODUCTS = 24,
+	OBJECTCATEGORY_GOLD = 30,
+	OBJECTCATEGORY_DEFAULT = 31, // unassigned loot
+
+	OBJECTCATEGORY_FIRST = OBJECTCATEGORY_ARMORS,
+	OBJECTCATEGORY_LAST = OBJECTCATEGORY_DEFAULT,
+};
