+++ src/game/movement/position.cpp
@@ -12,6 +12,12 @@
 #include "game/movement/position.hpp"
 #include "utils/tools.hpp"
 
+double Position::getEuclideanDistance(const Position &p1, const Position &p2) {
+	int32_t dx = Position::getDistanceX(p1, p2);
+	int32_t dy = Position::getDistanceY(p1, p2);
+	return std::sqrt(dx * dx + dy * dy);
+}
+
 Direction Position::getRandomDirection() {
 	static std::vector<Direction> dirList {
 		DIRECTION_NORTH,
