+++ src/items/item.cpp
@@ -28,6 +28,20 @@
 Items Item::items;
 
 std::shared_ptr<Item> Item::CreateItem(const uint16_t type, uint16_t count /*= 0*/, Position* itemPosition /*= nullptr*/) {
+	// A map which contains items that, when on creating, should be transformed to the default type.
+	static const phmap::flat_hash_map<ItemID_t, ItemID_t> ItemTransformationMap = {
+		{ ITEM_SWORD_RING_ACTIVATED, ITEM_SWORD_RING },
+		{ ITEM_CLUB_RING_ACTIVATED, ITEM_CLUB_RING },
+		{ ITEM_DWARVEN_RING_ACTIVATED, ITEM_DWARVEN_RING },
+		{ ITEM_RING_HEALING_ACTIVATED, ITEM_RING_HEALING },
+		{ ITEM_STEALTH_RING_ACTIVATED, ITEM_STEALTH_RING },
+		{ ITEM_TIME_RING_ACTIVATED, ITEM_TIME_RING },
+		{ ITEM_PAIR_SOFT_BOOTS_ACTIVATED, ITEM_PAIR_SOFT_BOOTS },
+		{ ITEM_DEATH_RING_ACTIVATED, ITEM_DEATH_RING },
+		{ ITEM_PRISMATIC_RING_ACTIVATED, ITEM_PRISMATIC_RING },
+		{ ITEM_OLD_DIAMOND_ARROW, ITEM_DIAMOND_ARROW },
+	};
+
 	std::shared_ptr<Item> newItem = nullptr;
 
 	const ItemType &it = Item::items[type];
