+++ src/kv/CMakeLists.txt
@@ -1,5 +1,6 @@
 target_sources(${PROJECT_NAME}_lib PRIVATE
     value_wrapper.cpp
+    value_wrapper_proto.cpp
     kv.cpp
     kv_sql.cpp
 )
