+++ src/kv/kv.hpp
@@ -9,11 +9,15 @@
 
 #pragma once
 
-#include <string>
-#include <mutex>
-#include <initializer_list>
-#include <parallel_hashmap/phmap.h>
-#include <optional>
+#ifndef USE_PRECOMPILED_HEADERS
+	#include <string>
+	#include <mutex>
+	#include <initializer_list>
+	#include <parallel_hashmap/phmap.h>
+	#include <optional>
+	#include <unordered_set>
+	#include <iomanip>
+#endif
 
 #include "lib/logging/logger.hpp"
 #include "kv/value_wrapper.hpp"
