+++ src/lib/logging/log_with_spd_log.hpp
@@ -18,9 +18,9 @@ public:
 	static Logger &getInstance();
 
 	void setLevel(const std::string &name) override;
-	[[nodiscard]] std::string getLevel() const override;
+	std::string getLevel() const override;
 
-	void log(std::string lvl, fmt::basic_string_view<char> msg) const override;
+	void log(const std::string &lvl, fmt::basic_string_view<char> msg) const override;
 };
 
 constexpr auto g_logger = LogWithSpdLog::getInstance;
