+++ src/lua/functions/creatures/combat/variant_functions.cpp
@@ -11,6 +11,7 @@
 
 #include "items/cylinder.hpp"
 #include "lua/functions/creatures/combat/variant_functions.hpp"
+#include "lua/global/lua_variant.hpp"
 
 int VariantFunctions::luaVariantCreate(lua_State* L) {
 	// Variant(number or string or position or thing)
