+++ src/lua/global/lua_timer_event_descr.hpp
@@ -1,0 +1,27 @@
+/**
+ * Canary - A free and open-source MMORPG server emulator
+ * Copyright (©) 2019-2022 OpenTibiaBR <opentibiabr@outlook.com>
+ * Repository: https://github.com/opentibiabr/canary
+ * License: https://github.com/opentibiabr/canary/blob/main/LICENSE
+ * Contributors: https://github.com/opentibiabr/canary/graphs/contributors
+ * Website: https://docs.opentibiabr.com/
+ */
+
+#pragma once
+
+#ifndef USE_PRECOMPILED_HEADERS
+	#include <cstdint>
+	#include <list>
+	#include <string>
+#endif
+
+struct LuaTimerEventDesc {
+	int32_t scriptId = -1;
+	std::string scriptName;
+	int32_t function = -1;
+	std::list<int32_t> parameters;
+	uint32_t eventId = 0;
+
+	LuaTimerEventDesc() = default;
+	LuaTimerEventDesc(LuaTimerEventDesc &&other) = default;
+};
