+++ src/lua/scripts/script_environment.hpp
@@ -74,7 +74,6 @@ public:
 	void removeItemByUID(uint32_t uid);
 
 private:
-	using VariantVector = std::vector<const LuaVariant*>;
 	using StorageMap = std::map<uint32_t, int32_t>;
 	using DBResultMap = std::map<uint32_t, DBResult_ptr>;
